Query data of an entity in a collection.

This functions queries data of a specified entitiy in a specified collection.

declass_entity_data(collection, entity, page_size = 25, page = 1,
  num_pages = 10)

Arguments

collection
The collection name

entity
The entity name

page_size
The number of rows displayed per page. Default setting is 25 records per page.

page
The starting page number to be queried. Default setting is page No.1.

num_pages
The number of pages to be quried. Default setting is 10 pages.

Value

Fields of data for the specified entity in that collection, and urls.

Examples

declass_entity_data("kissinger", "countries", page_size = 30, page = 1, num_pages = 5)

#> [1] "http://api.declassification-engine.org/declass/v0.4/entity_info/?collection=kissinger&entity=countries&page_size=30&page=1"
#> [1] "http://api.declassification-engine.org/declass/v0.4/entity_info/?collection=kissinger&entity=countries&page_size=30&page=2"
#> [1] "http://api.declassification-engine.org/declass/v0.4/entity_info/?collection=kissinger&entity=countries&page_size=30&page=3"
#> [1] "http://api.declassification-engine.org/declass/v0.4/entity_info/?collection=kissinger&entity=countries&page_size=30&page=4"
#> [1] "http://api.declassification-engine.org/declass/v0.4/entity_info/?collection=kissinger&entity=countries&page_size=30&page=5"

#>     count  id                                     name
#> 1       5 170                                 Colombia
#> 2       3 174                                  Comoros
#> 3      21 180         Democratic Republic of the Congo
#> 4       5 188                               Costa Rica
#> 5      67 192                                     Cuba
#> 6     142 196                                   Cyprus
#> 7       3 200                           Czechoslovakia
#> 8       1 204                                    Benin
#> 9      22 208                                  Denmark
#> 10      5 214                       Dominican Republic
#> 11     15 218                                  Ecuador
#> 12      4 230                                 Ethiopia
#> 13      1 242                                     Fiji
#> 14     23 246                                  Finland
#> 15    258 250                                   France
#> 16      6 268                        Georgia (country)
#> 17     24 275                       State of Palestine
#> 18     46 278                             East Germany
#> 19     48 280                             West Germany
#> 20      4 288                                    Ghana
#> 21    229 300                                   Greece
#> 22      3 316                                     Guam
#> 23      5 320                                Guatemala
#> 24      6 324                                   Guinea
#> 25     10 328                                   Guyana
#> 26      2 332                                    Haiti
#> 27      4 336                             Vatican City
#> 28      1 340                                 Honduras
#> 29      2 344                                Hong Kong
#> 30      3 348                                  Hungary
#> 31      4 352                                  Iceland
#> 32     54 356                                    India
#> 33     15 360                                Indonesia
#> 34     74 364                                     Iran
#> 35     19 368                                     Iraq
#> 36      4 372                      Republic of Ireland
#> 37    868 376                                   Israel
#> 38     78 380                                    Italy
#> 39      8 388                                  Jamaica
#> 40     97 392                                    Japan
#> 41     88 400                                   Jordan
#> 42     22 404                                    Kenya
#> 43      7 410                              South Korea
#> 44     13 414                                   Kuwait
#> 45     15 418                                     Laos
#> 46     83 422                                  Lebanon
#> 47      4 430                                  Liberia
#> 48     25 434                                    Libya
#> 49      1 440                                Lithuania
#> 50      4 442                               Luxembourg
#> 51      2 458                                 Malaysia
#> 52      4 462                                 Maldives
#> 53      4 466                                     Mali
#> 54      1 470                                    Malta
#> 55      6 474                               Martinique
#> 56      4 478                               Mauritania
#> 57     72 484                                   Mexico
#> 58      1 496                                 Mongolia
#> 59      2 499                               Montenegro
#> 60     12 504                                  Morocco
#> 61      9 508                               Mozambique
#> 62      3 512                                     Oman
#> 63      7 516                                  Namibia
#> 64      1 524                                    Nepal
#> 65     16 528                              Netherlands
#> 66      7 554                              New Zealand
#> 67      3 558                                Nicaragua
#> 68      6 562                                    Niger
#> 69      6 566                                  Nigeria
#> 70      3 574                           Norfolk Island
#> 71     11 578                                   Norway
#> 72     37 586                                 Pakistan
#> 73      8 594                        Panama Canal Zone
#> 74      1 598                         Papua New Guinea
#> 75      1 600                                 Paraguay
#> 76     10 604                                     Peru
#> 77     10 608                              Philippines
#> 78     18 616                                   Poland
#> 79     38 620                                 Portugal
#> 80      4 626                               East Timor
#> 81     17 630                              Puerto Rico
#> 82      1 634                                    Qatar
#> 83      1 638                                  Reunion
#> 84     28 642                                  Romania
#> 85      1 659                    Saint Kitts and Nevis
#> 86      1 660                                 Anguilla
#> 87      5 662                              Saint Lucia
#> 88      2 678                  SAPSo Tome and PrAncipe
#> 89     72 682                             Saudi Arabia
#> 90      6 686                                  Senegal
#> 91      4 688                                   Serbia
#> 92      1 690                               Seychelles
#> 93      3 702                                Singapore
#> 94    166 704                                  Vietnam
#> 95      6 706                                  Somalia
#> 96     60 710                             South Africa
#> 97     27 716                                 Zimbabwe
#> 98     26 724                                    Spain
#> 99     18 728                              South Sudan
#> 100     1 729                                    Sudan
#> 101     9 748                                Swaziland
#> 102    17 752                                   Sweden
#> 103   101 756                              Switzerland
#> 104   439 760                                    Syria
#> 105    31 764                                 Thailand
#> 106     1 780                      Trinidad and Tobago
#> 107     2 784                     United Arab Emirates
#> 108     5 788                                  Tunisia
#> 109   280 792                                   Turkey
#> 110     1 795                             Turkmenistan
#> 111     6 800                                   Uganda
#> 112     2 804                                  Ukraine
#> 113   910 810                             Soviet Union
#> 114   577 818                                    Egypt
#> 115   366 826                           United Kingdom
#> 116     9 834                                 Tanzania
#> 117   937 840                            United States
#> 118     2 854                             Burkina Faso
#> 119     1 858                                  Uruguay
#> 120    29 862                                Venezuela
#> 121     2 887                                    Yemen
#> 122    26 890 Socialist Federal Republic of Yugoslavia
#> 123     2 891                    Serbia and Montenegro
#> 124    22 894                                   Zambia