Query document by ID.

This function queries a specific document or documents by user-inputed ID. It is intended for users who know what they want and have specific IDs in mind.

declass_query_id_data(id_list)

Arguments

id_list
A list of IDs

Value

Content of the entire document

Examples

declass_query_id_data("Clinton-74931")

#> $count
#> [1] 1
#> 
#> $results
#>   body_summary cable_references classification cable_type topics from_field
#> 1           NA               NA   UNCLASSIFIED         NA     NA         NA
#>              id               subject category office                 title
#> 1 Clinton-74931  Fwd: From Joe Wilson       NA     NA  Fwd: From Joe Wilson
#>   nuclear to_field source location type refs
#> 1      NA       NA     NA       NA   NA   NA
#>                                                                                                                                                                                                                                                                                                                body
#> 1 From:                              sbwhoeop\nSent:                              Tuesday, November 10, 2009 11:47 AM\nTo:\nSubject:                           Fwd: From Joe Wilson\n\n\nFYI. How did it go in Berlin? Looked terrific. What does Merkel think of the Blair option? Sid\n\n\n    Original Message\n
#>   description tags date_year body_html collection            persons
#> 1          NA   NA        NA        NA    clinton Blumenthal, Sidney
#>                  date readability chapt_title countries date_month source_path
#> 1 2009-11-10T11:47:00          NA          NA        NA         NA          NA
#>   title_docview concepts
#> 1            NA       NA
#>                                                                                                      pdf
#> 1 http://source.history-lab.org/clinton/source/pdf/HRCEmail_JulyWeb_Web_047_DOC_0C05765644_C05765644.pdf
#>   orighand
#> 1       NA
#> 

declass_query_id_data("1977LONDON17830,1975MOSCOW15238,P790162-1195")

#> $count
#> [1] 3
#> 
#> $results
#>   body_summary cable_references classification cable_type topics
#> 1           NA               NA   CONFIDENTIAL       full     NA
#> 2           NA               NA           <NA>  withdrawn     NA
#> 3           NA               NA         SECRET      preel     NA
#>              from_field              id                                subject
#> 1                MOSCOW 1975MOSCOW15238 KAPITSA ON SOUTHEAST ASIA CONFIDENTIAL
#> 2                LONDON 1977LONDON17830                                   <NA>
#> 3 NEA SS TARNOFF, PETER    P790162-1195        ARAB STATES SUMMIT PREPARATIONS
#>   category     office                                  title nuclear
#> 1       NA ACTION EUR KAPITSA ON SOUTHEAST ASIA CONFIDENTIAL      NA
#> 2       NA       <NA>                                   <NA>      NA
#> 3       NA ORIGIN NEA        ARAB STATES SUMMIT PREPARATIONS      NA
#>                           to_field source location type refs
#> 1                            STATE     NA       NA   TE   NA
#> 2                            STATE     NA       NA <NA>   NA
#> 3 WHITE HOUSE BRZEZINSKI, ZBIGNIEW     NA       NA   ME   NA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     body
#> 1 1.  SUMMARY.  IN CONVERSATION WITH KAPITSA OCTOBER 22 (REFTEL)\nHE DISCOUNTED SIHANOUK'S INFLUENCE IN CAMBODIA, AND SAID THAT THE\nCHINESE CURRENTLY HAVE THE GREATEST VOICE IN THAT COUNTRY.  OVER\nTHE LONG RUN, HOWEVER, HE EXPECTED THAT THE CHINESE WOULD LOSE\nTHEIR DOMINANT INFLUENCE IN PHNOM PENH.  IN VIETNAM, HE SAID THAT\nTHE REUNIFICATION PROCESS HAS FOR PRACTICAL PURPOSES ALMOST TAKEN\nPLACE, AND THAT THE CURRENT POLITICAL CLIMATE IS NOT CONDUCIVE\nTO RESOLUTION OF U.S.-VIETNAMESE PROBLEMS.  HE DENIED THAT\nVIETNAM HAS ANY AGGRESSIVE INTENTIONS TOWARD NEIGHBORS SUCH AS\nTHAILAND.  ACCORDING TO KAPITSA, HANOI IS CLOSER TO THE USSR THAN\nCHINA ON WORLD ISSUES.  KAPITSA STRESSED THAT SOVIET COLLECTIVE\nSECURITY PROPOSALS WERE LONG-RANGE IN NATURE AND THAT SOVIETS DID\nNOT EXPECT ANY PROGRESS IN THIS REGARD DURING THE NEXT DECADE.  HE\nALSO POINTED OUT THAT THE USSR IS NOT HOSTILE TO ASEAN OR TO\nTHE ASEAN COUNTRIES.  END SUMMARY.\n2.  COMBODIA.  WHEN ASKED ABOUT THE CURRENT SITUATION IN CAMBODIA,\nKAPITSA COMMENTED THAT SIHANOUK HAS NO POLITICAL IMPORTANCE.  HE\nHAS NO FUTURE IN CAMBODIA, KAPITSA SAID.  IN SOME AREAS PRO-\nCHINESE ELEMENTS DOMINATE, IN OTHER PLACES PRO-VIETNAMESE ARE IN\nCHARGE, AND IN STILL OTHER LOCALITIES NATIONALISTS, WHO OPPOSE\nBOTH THE OTHER GROUPS, WIELD POWER.  THE KEY PROBLEM IN CAMBODIA\nNOW IS THE INTERNAL STRUGGLE BETWEEN THESE FACTIONS.  THOSE\nCAMBODIANS WHO WERE NOT PRESENT DURING THE WAR, KAPITSA SAID,\nMORE AND MORE ARE INFLUENCING EVENTS IN PHNOM PENH, BUT THEY\nDO NOT HAVE DEEP SUPPORT WITHIN THE COUNTRY.  HOWEVER, AFTER\nANALYZING THE STRENGTH OF THE THREE GROUPS HE IS CONFIDENT THAT\nTHE CHINESE WOULD EVENTUALLY LOSE THEIR DOMINANT INFLUENCE IN\nPHNOM PENH.  CAMBODIA WILL NOT BE TURNED INTO ANOTHER ALBANIA,\nKAPITSA SAID, PLACING THE ODDS AT 60/40 THAT THIS WOULD NOT OCCUR.\nHE SAID THE BASED THIS BELIEF ON THE NATURE OF THE CAMBODIAN\nPEOPLE AND THE INEVITABLE INFLUENCE OF THE VIETNAMESE.  KAPITSA\nALSO COMMENTED THAT NO FOREIGN EMBASSIES OTHER THAN VIETNAM AND\nCHINA WOULD BE PERMITTED IN PHNOMPENH FOR A LONG TIME.\n3.  VIETNAM. WHEN ASKED ABOUT PROSPECTS FOR REUNIFICATION OF THE\nTWO VIETNAMS, KAPITSA REPLIED THAT THE REUNIFICATION PROCESS HAS\nFOR ALL PRACTICAL EFFECTS ALREADY TAKEN PLACE, THAT THERE IS ONE\nCOMMUNIST PARTY IN CHARGE, AND THAT JOINT PLANNING FOR BOTH\nCOUNTIRES HAS NOW BEEN IMPLEMENTED.  FOR THIS REASON, HE SAID, HE\nANTICIPATES NO NEED FOR FOREIGN EMBASSIES IN SOUTH VIETNAM.  ON\nTHE OTHER HAND, KAPITSA ADDED, THE VIETNAMESE HAVE TO SOLVE SOCIAL,\nECONOMIC, AND POLITICAL PROBLEMS IN SOUTH VIETNAM BEFORE ACTUAL\nJURIDICAL REUNIFICATION CAN TAKE PLACE.  CHOLON, FOR EXAMPLE, IS A\nGREAT PROBLEM BECAUSE OF THE 1.2 MILLION CHINESE THERE WHO\nACCUMULATED HUGE WEALTH DURING THE WAR.  FOR THE VIETNAMESE THIS\nIS NOT AN EASY PROBLEM, ESPECIALLY AS THE CHINESE HAVE DEMONSTRA-\nTED AN INTEREST IN THE POPULATION OF CHOLON.  THE VIETNAMESE\nATTITUDE, HOWEVER, IS THAT IT IS NONE OF PEKING'S BUSINESS.\n"THIEU" HAD DECLARED THEM ALL VIETNAMESE CITIZENS AND THE VIET-\nNAMESE NOW ACCEPT THIS.  THE CAPITAL ACCUMULATED BY THE CHINESE\nIS ALSO NONE OF PEKING'S BUSINESS, ACCORDING TO HANOI.  THE\nOVERSEAS CHINESE IN VIETNAM HAVE NOW BEEN DEPRIVED OF THEIR\nMONEY THROUGH RECENTLY IMPLEMENTED FINANCIAL REFORMS, IN ANY\nCASE, BUT THIS STILL REMAINS A PROBLEM BETWEEN HANOI AND PEKING.\nON THE OTHER HAND, KAPITSA SAID IT WAS HIS IMPRESSION THAT CHINA\nEXAGGERATES THE DIFFERENCES BETWEEN THE VIETNAMESE AND THEM-\nSELVES IN ACCUSING HANOI OF HOSTILE INTENTIONS TOWARD ITS\nNEIGHBORS.  VIETNAM HAS SO MANY INTERNAL PROBLEMS TO SOLVE IT\nIS "STRANGE" TO ACCUSE HANOI OF SUCH INTENTIONS TOWARD NEIGHBORS\nLIKE THAILAND, KAPITSA SAID.\n4.  IF THE TWO VIETNAMS ARE CLOSE TO REUNIFICATION, I SAID, THIS\nWOULD SURELY TAKE AWAY SOME OF THE FORCE FROM THE SOUTH VIET-\nNAMESE CLAIM TO BE REPRESENTED AS A SEPARATE ENTITY IN THE UN.\nSOMEWHAT DISCOMFITED, KAPITSA REPLIED THAT THE SOUTH VIETNAMESE\nHAVE THIS RIGHT.  THE SOCIAL DEVELOPMENT OF THE NORTH AND SOUTH\nARE NOT AT THE SAME LEVEL, HE SAID.  THE SOUTH IS STILL A\nBOURGEOIS, DEMOCRATIC STATE AND HAS A DIFFERENT ECONOMIC, SOCIAL\nAND POLITICAL STRUCTURE THAN THE NORTH.  THE SOUTH VIETNAMESE\nGOVERNMENT IS COMPOSED OF PROGRESSIVES, SAIGON REGIME ELEMENTS,\nAND ALSO THE SO-CALLED THIRD FORCES.  HOWEVER, KAPITSA ADDED,\nMANY OF THE LATTER WAITED IN FRANCE FOR TOO LONG REALLY TO\nINFLUENCE FUTURE DEVELOPMENTS IN THE SOUTH.  KAPITSA STATED THAT\nHE UNDERSTOOD OUR OBJECTIONS AGAINST SEATING BOTH VIETNAMS IN\nTHE UN, BUT SAID THAT IT WAS HIS IMPRESSION THAT THE U.S. IS THE\nULTIMATE LOSER BECAUSE OF THIS POLICY AND WILL BE ISOLATED IN\nPURSUING IT.\n5.  I SAID TO KAPITSA THAT WE WELCOME THE FACT THAT HANOI HAS\nINDICATED WILLINGNESS TO LET SOME AMERICANS BE REPATRIATED FROM\nVIETNAM (STATE 250584) BUT ADDED THAT THERE ARE STILL OTHER\nAMERICAN DETAINED IN VIETNAM WHOM WE ARE ANXIOUS ABOUT.  WE ALSO\nHAVE NOT AS YET RECEIVED SATISFACTORY INFORMATION FROM THE VIETNAM-\nESE ABOUT AMERIANS MISSING IN ACTION DURING THE WAR.  MAKING CLEAR\nTHAT I WAS NOT SPEAKING ON THE BASIS OF INSTRUCTIONS, I ASKED\nKAPITSA WHETHER THIS PROBLEM MIGHT BE DISCUSSED WITH LE DUAN\nDURING HIS VISIT HERE THE END OF THIS MONTH.  (KAPITSA HAD\nMENTIONED SUCH A POSSIBILITY TO AUSTRALIAN AMBASSADOR PLIMSOLL\nIN A RECENT CONVERSATION--REFTEL A).  KAPITSA REPLIED THAT IT\nWAS THE U.S. WHICH CREATED A PROBLEM IN THIS REGARD AND THAT\nTHERE DOES NOT EXIST THE RIGHT POLITICAL ATMOSPHERE FOR MAKING\nPROGRESS ON THESE ISSUES.  HE SAID BY FOLLOWING POLICIES SUCH AS\nTHE VETOING OF THE VIETNAMESE FROM ENTRY INTO THE UN WASHINGTON IS\nJUST ISOLATING ITSELF.  "WE DO NOT WANT YOU TO BE ISOLATED,"\nKAPITSA ADDED, "WE HAVE THE SAME GENERAL INTERESTS THAT YOU HAVE."\n(DESPITE KAPITSA'S GENERALLY NEGATIVE REPLY, IT WAS OF INTEREST\nTHAT IGOR ROGACHEV, KAPITSA'S DEPUTY, WHO WAS ALSO PRESENT AS WAS\nPOLITICAL CONSELOR BREMENT, MADE A CAREFUL NOTE WHEN I RAISED\nTHE POSSIBLITIY OF MENTIONING THIS SUBJECT TO LE DUAN.)\nFM AMEMBASSY MOSCOW\nTO SECSTATE WASHDC 5940\nINFO AMEMBASSY BANGKOK\nAMEMBASSY BELGRADE\nAMEMBASSY BERLIN\nUSMISSION USBERLIN\nAMEMBASSY BUCHAREST\nAMEMBASSY BUDAPEST\nCINCPAC\nAMCONSUL HONG KONG\nAMEMBASSY JAKARTA\nAMEMBASSY KUALA LUMPUR\nAMCONSUL LENINGRAD\nAMEMBASSY LONDON\nAMEMBASSY MANILA\nAMCONSUL MUNICH\nAMEMBASSY PARIS\nAMEMBASSY PRAGUE\nAMEMBASSY SINGAPORE\nAMEMBASSY SOFIA\nAMEMBASSY TAIPEI\nAMEMBASSY VIENTIANE\nAMEMBASSY WARSAW\nC O N F I D E N T I A L SECTION 2 OF 2 MOSCOW 15238\nCINCPAC FOR POLAD\n6.  WHEN ASKED FURTHER ABOUT THE LE DUAN VISIT, KAPITSA SAID THAT\nTHE VIETNAMESE FIRST SECRETARY WILL BE COMING HERE FOR TALKS AT\nTHE END OF THE MONTH AND THAT AFTER THE TALKS HE WILL THEN\nVACATION IN THE USSR.  THE VIETNAMESE NEED MUCH AID FROM US, HE\nSAID, INCLUDING GRAIN.  HE POINTED OUT IN THIS REGARD THAT\nALTHOUGH SOME DIFFERENCES HAD EMERGED DURING RECENT VIETNAMESE\nCONVERSATIONS WITH THE CHINESE, PEKING DID GIVE VIET NAM GENEROUS\nASSISTANCE AS A RESULT OF THOSE TALKS.\n7.  HANOI HAS A DIFFICULT PROBLEM IN BALANCING OFF RELATIONS WITH\nTHE SOVIET UNION AND WITH CHINA, KAPITSA SAID.  THE VIETNAMESE\nSYMPATHIZE WITH THE USSR GENERALLY AND DO NOT AGREE WITH THE\nCHINESE ASSESSMENT THAT THE USSR IS A SOCIAL-IMPERIALIST FASCIST\nSTATE.  AT THE SAME TIME CHINA IS A HUGE NEIGHBOR AND ITS VIEWS\nMUST BE CONSIDERED BY THE VIETNAMESE.  THE VIETNAMESE WILL,\nTHEREFORE, CONTINUE THE SAME POLICY THAT THEY CARRIED ON DURING\nTHE WAR IN THEIR CONDUCT OF RELATIONS WITH MOSCOW AND PEKING.\nIT WILL GO BACK AND FORTH, SAID KAPITSA, SOMETIME IT WILL BE\n60/40 AND SOMETIME 50/50, BUT GENERALLY THE VIETNAMESE ARE\nCLOSER TO THE USSR THAN TO CHINA ON VARIOUS WORLD ISSUES.\n8.  ASIAN COLLECTIVE SECURITY. IN RESPONSE TO A QUESTION AS TO\nWHETHER THE VIETNAMESE SUPPORT SOVIET PROPOSALS FOR A COLLECTIVE\nSECURITY AGREEMENT IN ASIA, KAPITSA SAID THAT DURING THE WAR\nHANOI WAS COMPLETELY AGAINST THIS IDEA.  NOW, HOEVER, THEY ARE NOT\nAGAINST IT FORMALLY AND DEFINITELY, BUT ARE ALSO NOT IN A HURRY\nTO SUPPORT IT.  KAPITSA SAID THAT IN ANY CASE THE SOVIET COLLECTIVE\nSECURITY PROPOSALS WERE LONG-RANGE AND MIGHT NOT BE IMPLEMENTED\nFOR PERHAPS FIFTY YEARS, CERTAINLY NOT WITHIN THE NEXT FIVE OR\nTEN YEARS.  FURTHERMORE, A FORMULA ON ASIAN SECURITY CANNOT BE\nFULLY ARTICULATED AT THIS TIME.  WHAT WE ENVISION, HE SAID, MIGHT\nBE A TREATY OF PEACE OR OF MUTUAL SECURITY, AN AGREEMENT ON THE\nNON-USE OF FORCE OR AGAINST AGRESSION, OR SOMETHING ALONG SUCH\nLINES.  WHEN ASKED AS TO WHAT AREA SUCH AN AGREEMENT WOULD COVER,\nKAPITSA SAID THAT IT WOULD CERTAINLY INCLUDE THE MIDDLE EAST AND\nPROBABLY AUSTRALIA AND NEW ZEALAND, AND THAT THE MAJOR POWERS\nSHOULD ALSO BE INCLUDED.  THE FIRST OBLIGATION WOULD BE TO\nREFRAIN FROM WAR AS AN INSTRUMENT OF POLICY; IN OTHER WORDS,\nSUPPORT FOR THE NON-USE OF FORCE AND NON-INTERFERENCE IN INTERNAL\nAFFAIRS OF OTHER COUNTRIES.  WE ARE PRACTICAL PEOPLE, HE SAID,\nAND REALIZE THAT THE SETTLEMENT OF BORDER PROBLEMS WILL BE FAR\nMORE DIFFICULT IN ASIA THAN IN EUROPE.  THUS ANY STATEMENT OF\nPRINCIPLE ON THE INVIOLABILITY OF BORDERS MIGHT NOT BE ACCEPTED.\n9.  WHAT WE ARE DOING NOW, KAPITSA CONTINUED, IS COMPARABLE TO\nA HUNTER IN THE WOODS WHO HAS A FIRE GOING.  EVERY ONCE IN A\nWHILE WE ADD A LOG TO THE FIRE TO KEEP IT BURNING, BUT WE ARE NOT\nIN A HURRY AND WE HAVE NO PLANS TO ACCELERATE OUR EFFORTS.\nFURTHERMORE, WE REALIZE THAT WHEREAS SUVERAL YEARS AGO 60 PER\nCENT OF THE ASIAN STATES SUPPORTED THIS GENERAL CONCEPT, NOW ONLY\n40 PER CENT FAVOR IT.  KAPITSA REITERATED THAT THIS SOVIET\nPROPOSAL WAS NOT DIRECTED AGAINST CHINA.  WE DO NOT WISH TO SET\nUP SUCH AN AGREEMENT WITHOUT CHINA.  THIS WOULD CONSTITUTE A BLOC\nAND BLOCS ARE NOW OUTMODED.\n10.  ASEAN. WHEN ASKED ABOUT RECENT SOVIET PRESS ARTICLES WHICH\nHAVE BEEN FAIRLY HOSTILE TO ASEAN, KAPITSA REPLIED THAT ONE HAD\nTO DIFFERENTIATE BETWEEN THE SOVIET PROPAGANDA POSITION AND ITS\nDIPLOMATIC POSITION.  WE ARE REALISTS, HE STATED, AND CARRY ON\nA REALISTIC POLICY.  OCCASIONALLY, ARTICLES APPEAR IN THE PRESS,\nBUT THIS IS LARGELY THE RESULT OF INERTIA AND, HE SAID, WE ARE\nCERTAINLY NOT HOSTILE TO ASEAN OR THE ASEAN COUNTRIES.\nSTOESSEL\n
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   <NA>
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#>   description tags date_year body_html      collection persons       date
#> 1          NA   NA        NA        NA statedeptcables      NA 1975-10-23
#> 2          NA   NA        NA        NA statedeptcables      NA 1977-10-28
#> 3          NA   NA        NA        NA statedeptcables      NA 1979-10-17
#>   readability chapt_title                  countries date_month
#> 1          NA          NA China, Japan, Soviet Union         NA
#> 2          NA          NA                       <NA>         NA
#> 3          NA          NA               Saudi Arabia         NA
#>                                                                             source_path
#> 1              http://source.history-lab.org/statedeptcables/source/xml/1/TEL.OCT75.RPU
#> 2            http://source.history-lab.org/statedeptcables/source/xml/3/TEL.WD.OCT77.PU
#> 3 http://source.history-lab.org/statedeptcables/source/xml/6/1979-P-REEL-Bucket1-10.xml
#>   title_docview
#> 1            NA
#> 2            NA
#> 3            NA
#>                                                                                                           concepts
#> 1 FOREIGN RELATIONS, COLLECTIVE SECURITY, FOREIGN POLICY POSITION,  REGIONAL ORGANIZATIONS, DIPLOMATIC DISCUSSIONS
#> 2                                                                                                              n/a
#> 3                                                                              ARABS, SUMMIT MEETINGS, PEACE TALKS
#>   pdf orighand
#> 1  NA      n/a
#> 2  NA     <NA>
#> 3  NA      n/a
#>